requirements.txt
bitbucket-api==0.5.0
Flask==0.10.1
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
oauthlib==1.0.3
requests==2.8.1
requests-oauthlib==0.5.0
sh==1.11
todoist-python==0.2.20
Werkzeug==0.11.2
wheel==0.24.0